+++ src/Umbraco.Core/Persistence/Repositories/TagsRepository.cs
@@ -137,8 +137,6 @@ namespace Umbraco.Core.Persistence.Repositories
         {
             ((Entity)entity).AddingEntity();
 
-            entity.SanitizeTagsForXmlStorage();
-
             var factory = new TagFactory();
             var dto = factory.BuildDto(entity);
