+++ src/Umbraco.Core/Umbraco.Core.csproj
@@ -291,7 +291,6 @@
     <Compile Include="Dictionary\ICultureDictionaryFactory.cs" />
     <Compile Include="Dynamics\DynamicInstanceHelper.cs" />
     <Compile Include="Dynamics\DynamicXmlConverter.cs" />
-    <Compile Include="EntityExtensions.cs" />
     <Compile Include="Events\CancellableEventArgs.cs" />
     <Compile Include="Events\ContentCacheEventArgs.cs" />
     <Compile Include="Events\CopyEventArgs.cs" />
